// src/types/partials/RelatedArtifact.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_types so do not edit manually

import { TExtension } from '../partials/Extension';
import { TMarkdown } from '../simpleTypes/Markdown';
import { TUrl } from '../simpleTypes/Url';
import { TAttachment } from '../partials/Attachment';
import { TCanonical } from '../simpleTypes/Canonical';

export type TRelatedArtifact = {
    id?: String;
    extension?: TExtension[];
    type: String;
    label?: String;
    display?: String;
    citation?: TMarkdown;
    url?: TUrl;
    document?: TAttachment;
    resource?: TCanonical;
};